add interactions

+++ Assets/Scripts/Configs/Mobs/BrainrotMobConfig.cs
@@ -7,6 +7,6 @@ public class BrainrotMobConfig : ScriptableObject
 {
     [field: SerializeField] public string Name { get; private set; }
     [field: SerializeField] public int BaseCost { get; private set; }
-    [field: SerializeField] public float ValuePerSecond { get; private set; }
+    [field: SerializeField] public int ValuePerSecond { get; private set; }
     [field: SerializeField] public GameObject MobPrefab { get; private set; }
 }
